move code into submodules
